# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bidly API Documentation",
        "description": "행사 기획 대행사와 용역사를 연결하는 입찰 플랫폼 API",
        "contact": {
            "name": "Bidly Support Team",
            "email": "support@bidly.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/projects/{project}/assign-evaluators": {
            "post": {
                "tags": [
                    "Evaluator Assignment"
                ],
                "summary": "프로젝트 전체 심사위원 배정",
                "description": "프로젝트와 연결된 모든 현재 및 미래 공고에 심사위원을 배정합니다.",
                "operationId": "8a08852665230256981d77c593998f0a",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "프로젝트 ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "evaluator_user_ids",
                                    "assignment_type"
                                ],
                                "properties": {
                                    "evaluator_user_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "example": [
                                            "01234567-89ab-cdef-0123-456789abcdef",
                                            "fedcba98-7654-3210-fedc-ba9876543210"
                                        ]
                                    },
                                    "assignment_type": {
                                        "type": "string",
                                        "enum": [
                                            "designated",
                                            "random"
                                        ],
                                        "example": "designated"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "심사위원 배정 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "프로젝트 전체에 심사위원이 성공적으로 배정되었습니다."
                                        },
                                        "assigned_evaluators_count": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "affected_announcements_count": {
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "권한 없음"
                    },
                    "404": {
                        "description": "프로젝트를 찾을 수 없음"
                    },
                    "422": {
                        "description": "유효성 검사 실패"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/rfps/{rfp}/assign-evaluators": {
            "post": {
                "tags": [
                    "Evaluator Assignment"
                ],
                "summary": "RFP 전체 심사위원 배정",
                "description": "RFP와 연결된 모든 현재 및 미래 공고에 심사위원을 배정합니다.",
                "operationId": "e5f4048ec85e8fd9452c87ae16128695",
                "parameters": [
                    {
                        "name": "rfp",
                        "in": "path",
                        "description": "RFP ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "evaluator_user_ids",
                                    "assignment_type"
                                ],
                                "properties": {
                                    "evaluator_user_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "example": [
                                            "01234567-89ab-cdef-0123-456789abcdef",
                                            "fedcba98-7654-3210-fedc-ba9876543210"
                                        ]
                                    },
                                    "assignment_type": {
                                        "type": "string",
                                        "enum": [
                                            "designated",
                                            "random"
                                        ],
                                        "example": "designated"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "심사위원 배정 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "RFP 전체에 심사위원이 성공적으로 배정되었습니다."
                                        },
                                        "assigned_evaluators_count": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "affected_announcements_count": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "권한 없음"
                    },
                    "404": {
                        "description": "RFP를 찾을 수 없음"
                    },
                    "422": {
                        "description": "유효성 검사 실패"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/announcements/{announcement}/assign-evaluators": {
            "post": {
                "tags": [
                    "Evaluator Assignment"
                ],
                "summary": "개별 공고 심사위원 배정",
                "description": "특정 공고에만 심사위원을 배정합니다.",
                "operationId": "a2d68478b8f91c12009c639fc21b6e12",
                "parameters": [
                    {
                        "name": "announcement",
                        "in": "path",
                        "description": "공고 ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "evaluator_user_ids",
                                    "assignment_type"
                                ],
                                "properties": {
                                    "evaluator_user_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "example": [
                                            "01234567-89ab-cdef-0123-456789abcdef",
                                            "fedcba98-7654-3210-fedc-ba9876543210"
                                        ]
                                    },
                                    "assignment_type": {
                                        "type": "string",
                                        "enum": [
                                            "designated",
                                            "random"
                                        ],
                                        "example": "designated"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "심사위원 배정 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "공고에 심사위원이 성공적으로 배정되었습니다."
                                        },
                                        "assigned_evaluators_count": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "권한 없음"
                    },
                    "404": {
                        "description": "공고를 찾을 수 없음"
                    },
                    "422": {
                        "description": "유효성 검사 실패"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}